server/requirements.txt
openenv-core
fastapi
uvicorn
scikit-learn
numpy
pydantic
openai